chore: 폴더 구조 변경 및 네비게이션 바 스타일 관련 파일 생성
