+++ package.json
@@ -18,6 +18,7 @@
     "@reduxjs/toolkit": "^2.3.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-icons": "^5.3.0",
     "react-redux": "^9.1.2",
     "react-router-dom": "^7.0.1",
     "styled-components": "^6.1.13",
